Move api requests to helper functions
