+++ client/src/styles/ratingsReviews.scss
@@ -84,9 +84,10 @@
     #review-dropdown {
       grid-row: 1 / 2;
 
-      h7 {
+      h6 {
         font-weight: bold;
         margin-right: 5px;
+        display: inline-flex;
       }
 
       select {
